requirements.txt
affinegap==1.10
ago==0.0.9
appdirs==1.4.3
beautifulsoup4==4.6.0
boto==2.47.0
boto3==1.4.4
botocore==1.5.58
bs4==0.0.1
BTrees==4.4.1
categorical-distance==1.9
certifi==2017.4.17
chardet==3.0.3
click==6.7
datetime-distance==0.1.3
dedupe==1.7.9
dedupe-hcluster==0.3.8
dedupe-variable-datetime==0.1.5
docutils==0.13.1
DoubleMetaphone==0.1
fastcluster==1.1.24
Flask==0.12.2
future==0.16.0
haversine==2.5.1
highered==0.2.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
Levenshtein-search==1.4.4
MarkupSafe==1.1.1
neo4j==4.4.0
numpy==1.20.2
packaging==16.8
persistent==4.2.4.2
pyhacrf-datamade==0.2.5
PyLBFGS==0.2.0.13
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2018.9
requests==2.17.3
rlr==2.4.3
s3transfer==0.1.10
simplecosine==1.2
simplejson==3.11.1
six==1.11.0
tabulate==0.8.2
urllib3==1.21.1
user-agent==0.1.9
Werkzeug==0.12.2
zope.index==4.4.0
zope.interface==4.7.2
humanize==3.12.0
